refactoring/issue#132: Apply structure suggested by @denproc
