+++ piq/utils/__init__.py
@@ -1,0 +1,7 @@
+from piq.utils.common import _adjust_dimensions, _validate_input, _validate_features
+
+__all__ = [
+    "_adjust_dimensions",
+    "_validate_input",
+    "_validate_features",
+]
