+++ piq/utils/common.py
